backend/requirements.txt
uvicorn
fastapi
transformers
groq
python-dotenv
websockets
langchain>=0.3.21
langchain_community>=0.3.20
sentence-transformers
langchain_huggingface
langchain-core>=0.3.48
faiss-cpu
